some translations added...
...hopefully many more to come soon.
I had not actually planned on doing any of this yet, but I was avoiding booting up Photoshop.
I am doing them a little out of order, and there are a small percentage I am not completely confident with the translation out of context. Those are subject to change at a future point if required.

Other note: 
I have decided we are going to keep using the term Kai for 改 as it is already the standard in the community, as well as using Kai Ni for 改二. That means no changing Kai to Revision, and no changing Ni to 2.